Schrodinger's weather
