+++ src/components/Settings.js
@@ -1,3 +1,10 @@
 export const settings = {
-    weatherKey : "111b9dbafb53b561b5efd23e53016a89"
+    weatherKey : "7d850121780539ee5826a65db9f69fa5",
+    nashville : {
+        lat: 36.1627,
+        long: -86.7816
+    }
 }
+
+// 7d850121780539ee5826a65db9f69fa5
+// f660d54afdf04b8a4b8020cd8e357035
